requirements.txt
pytest==7.4.3
Appium-Python-Client==3.1.1
selenium==4.14.0
